# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Device Service API
  version: 1.0.0
  description: |
    API for managing devices. A device has an `id`, `name`, and `state`.
    The `state` is an enum: `in-use` or `terminated`.

servers:
  - url: https://api.example.com
    description: Production
  - url: https://sandbox.api.example.com
    description: Sandbox

tags:
  - name: Devices
    description: Endpoints for creating and retrieving devices

paths:
  /devices:
    get:
      tags: [Devices]
      summary: List devices
      description: Retrieve all devices.
      operationId: listDevices
      responses:
        '200':
          description: A list of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
              examples:
                sample:
                  value:
                    - id: "23456"
                      name: "Gateway-01"
                      state: "in-use"
                    - id: "12345"
                      name: "Sensor-22"
                      state: "terminated"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Devices]
      summary: Create a new device
      description: Create a new device.
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
            examples:
              valid:
                value:
                  name: "Thermostat-7"
                  state: "in-use"
      responses:
        '201':
          description: Device created
          headers:
            Location:
              description: URL of the newly created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
              examples:
                created:
                  value:
                    id: "23456"
                    name: "Thermostat-7"
                    state: "in-use"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /devices/{id}:
    get:
      tags: [Devices]
      summary: Get device by id
      description: Retrieve a single device by its unique identifier.
      operationId: getDeviceById
      parameters:
        - name: id
          in: path
          required: true
          description: The device identifier
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The requested device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
              examples:
                sample:
                  value:
                    id: "23456"
                    name: "Gateway-01"
                    state: "in-use"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /devices/by-brand/{brand}:
    get:
      tags: [ Devices ]
      summary: List devices by brand
      description: Convenience endpoint to list all devices of a given brand.
      operationId: listDevicesByBrand
      parameters:
        - name: brand
          in: path
          required: true
          description: Brand name to filter by (case-insensitive match recommended)
          schema:
            type: string
            minLength: 1
            maxLength: 200
      responses:
        '200':
          description: Devices matching the provided brand
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /devices/by-state/{state}:
    get:
      tags: [ Devices ]
      summary: List devices by state
      description: Convenience endpoint to list all devices in a given state.
      operationId: listDevicesByState
      parameters:
        - name: state
          in: path
          required: true
          description: Device state to filter by
          schema:
            $ref: '#/components/schemas/State'
      responses:
        '200':
          description: Devices matching the provided state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Device:
      type: object
      required: [id, name, brand, state]
      properties:
        id:
          type: integer
          format: int64
          example: "123456"
          description: Unique identifier for the device
        name:
          type: string
          example: "Gateway-01"
          description: Human-friendly device name
          minLength: 1
          maxLength: 200
        brand:
          type: string
          description: Manufacturer or brand of the device
          example: "Acme"
          minLength: 0
          maxLength: 200
        state:
          $ref: '#/components/schemas/State'

    DeviceCreateRequest:
      type: object
      required: [name, brand, state]
      properties:
        name:
          type: string
          example: "Thermostat-7"
          minLength: 1
          maxLength: 200
        brand:
          type: string
          example: "Pizer"
          minLength: 1
          maxLength: 200
        state:
          $ref: '#/components/schemas/State'

    State:
      type: string
      description: Device lifecycle state
      enum: [in-use, available, inactive]
      example: in-use

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: The request payload is invalid.
        details:
          type: object
          additionalProperties: true

  responses:
    BadRequest:
      description: The request was invalid or cannot be otherwise served.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingField:
              value:
                code: BAD_REQUEST
                message: "Field 'name' is required."
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                code: NOT_FOUND
                message: "Device with id '123' was not found."
    Conflict:
      description: A resource conflict occurred (e.g., duplicate name if constrained).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              value:
                code: CONFLICT
                message: "A device with this name already exists."
    InternalServerError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            serverError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: "Unexpected error. Please try again later."