# src/main/java/com/company/device/device/controller/DeviceController.java
package com.company.device.device.controller;

import com.company.device.device.autogenerated.api.DevicesApi;
import com.company.device.device.autogenerated.dto.DeviceCreateRequestDto;
import com.company.device.device.autogenerated.dto.DeviceDto;
import com.company.device.device.autogenerated.dto.StateDto;
import com.company.device.device.domain.Device;
import com.company.device.device.domain.enums.State;
import com.company.device.device.mapper.DeviceMapper;
import com.company.device.device.services.DeviceService;
import io.swagger.annotations.Api;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@Api(tags = "Devices")
@RestController
public class DeviceController implements DevicesApi {
	private final DeviceService deviceService;
	private final DeviceMapper deviceMapper;

	public DeviceController(DeviceService deviceService,
			DeviceMapper deviceMapper) {
		this.deviceService = deviceService;
		this.deviceMapper = deviceMapper;
	}

	@Override public ResponseEntity<DeviceDto> createDevice(DeviceCreateRequestDto deviceCreateRequestDto) {
    return null;
	}

	@Override public ResponseEntity<DeviceDto> getDeviceById(Long id) {
		return ResponseEntity.ok(deviceMapper.toDto(deviceService.getDeviceById(id)));
	}

	@Override
	public ResponseEntity<List<DeviceDto>> listDevices() {
		return ResponseEntity.ok(deviceMapper.toListDto(deviceService.getAll()));
	}

	@Override public ResponseEntity<List<DeviceDto>> listDevicesByBrand(String brand) {
		return ResponseEntity.ok(deviceMapper.toListDto(deviceService.getDeviceByBrand(brand)));
	}

	@Override public ResponseEntity<List<DeviceDto>> listDevicesByState(StateDto stateDto) {
		State state=deviceMapper.toState(stateDto);
		return ResponseEntity.ok(deviceMapper.toListDto(deviceService.getDeviceByState(state)));
	}

}