# target/generated-sources/openapi/src/main/java/com/company/device/device/autogenerated/api/DevicesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.company.device.device.autogenerated.api;

import com.company.device.device.autogenerated.dto.DeviceCreateRequestDto;
import com.company.device.device.autogenerated.dto.DeviceDto;
import com.company.device.device.autogenerated.dto.ErrorDto;
import com.company.device.device.autogenerated.dto.StateDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-22T17:33:26.712568800+01:00[Europe/Warsaw]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Devices", description = "Endpoints for creating and retrieving devices")
public interface DevicesApi {

    String PATH_CREATE_DEVICE = "/devices";
    /**
     * POST /devices : Create a new device
     * Create a new device.
     *
     * @param deviceCreateRequestDto  (required)
     * @return Device created (status code 201)
     *         or The request was invalid or cannot be otherwise served. (status code 400)
     *         or A resource conflict occurred (e.g., duplicate name if constrained). (status code 409)
     *         or An unexpected error occurred on the server. (status code 500)
     */
    @Operation(
        operationId = "createDevice",
        summary = "Create a new device",
        description = "Create a new device.",
        tags = { "Devices" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Device created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DeviceDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"id\":\"23456\",\"name\":\"Thermostat-7\",\"state\":\"in-use\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "400", description = "The request was invalid or cannot be otherwise served.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"BAD_REQUEST\",\"message\":\"Field 'name' is required.\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "409", description = "A resource conflict occurred (e.g., duplicate name if constrained).", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"CONFLICT\",\"message\":\"A device with this name already exists.\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred on the server.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"Unexpected error. Please try again later.\"}"
                    )
                })

            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DevicesApi.PATH_CREATE_DEVICE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<DeviceDto> createDevice(
        @Parameter(name = "DeviceCreateRequestDto", description = "", required = true) @Valid @RequestBody DeviceCreateRequestDto deviceCreateRequestDto
    );


    String PATH_GET_DEVICE_BY_ID = "/devices/{id}";
    /**
     * GET /devices/{id} : Get device by id
     * Retrieve a single device by its unique identifier.
     *
     * @param id The device identifier (required)
     * @return The requested device (status code 200)
     *         or The specified resource was not found. (status code 404)
     *         or An unexpected error occurred on the server. (status code 500)
     */
    @Operation(
        operationId = "getDeviceById",
        summary = "Get device by id",
        description = "Retrieve a single device by its unique identifier.",
        tags = { "Devices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The requested device", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DeviceDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"id\":\"23456\",\"name\":\"Gateway-01\",\"state\":\"in-use\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "404", description = "The specified resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"NOT_FOUND\",\"message\":\"Device with id '123' was not found.\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred on the server.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"Unexpected error. Please try again later.\"}"
                    )
                })

            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DevicesApi.PATH_GET_DEVICE_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<DeviceDto> getDeviceById(
        @NotNull @Parameter(name = "id", description = "The device identifier", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    String PATH_LIST_DEVICES = "/devices";
    /**
     * GET /devices : List devices
     * Retrieve all devices.
     *
     * @return A list of devices (status code 200)
     *         or An unexpected error occurred on the server. (status code 500)
     */
    @Operation(
        operationId = "listDevices",
        summary = "List devices",
        description = "Retrieve all devices.",
        tags = { "Devices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A list of devices", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DeviceDto.class)), examples = {
                    @ExampleObject(
                        name = "",
                        value = "[{\"id\":\"23456\",\"name\":\"Gateway-01\",\"state\":\"in-use\"},{\"id\":\"12345\",\"name\":\"Sensor-22\",\"state\":\"terminated\"}]"
                    )
                })

            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred on the server.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"Unexpected error. Please try again later.\"}"
                    )
                })

            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DevicesApi.PATH_LIST_DEVICES,
        produces = { "application/json" }
    )
    ResponseEntity<List<DeviceDto>> listDevices(
        
    );


    String PATH_LIST_DEVICES_BY_BRAND = "/devices/by-brand/{brand}";
    /**
     * GET /devices/by-brand/{brand} : List devices by brand
     * Convenience endpoint to list all devices of a given brand.
     *
     * @param brand Brand name to filter by (case-insensitive match recommended) (required)
     * @return Devices matching the provided brand (status code 200)
     *         or The request was invalid or cannot be otherwise served. (status code 400)
     *         or An unexpected error occurred on the server. (status code 500)
     */
    @Operation(
        operationId = "listDevicesByBrand",
        summary = "List devices by brand",
        description = "Convenience endpoint to list all devices of a given brand.",
        tags = { "Devices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Devices matching the provided brand", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DeviceDto.class)))
            }),
            @ApiResponse(responseCode = "400", description = "The request was invalid or cannot be otherwise served.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"BAD_REQUEST\",\"message\":\"Field 'name' is required.\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred on the server.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"Unexpected error. Please try again later.\"}"
                    )
                })

            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DevicesApi.PATH_LIST_DEVICES_BY_BRAND,
        produces = { "application/json" }
    )
    ResponseEntity<List<DeviceDto>> listDevicesByBrand(
        @NotNull @Size(min = 1, max = 200) @Parameter(name = "brand", description = "Brand name to filter by (case-insensitive match recommended)", required = true, in = ParameterIn.PATH) @PathVariable("brand") String brand
    );


    String PATH_LIST_DEVICES_BY_STATE = "/devices/by-state/{state}";
    /**
     * GET /devices/by-state/{state} : List devices by state
     * Convenience endpoint to list all devices in a given state.
     *
     * @param state Device state to filter by (required)
     * @return Devices matching the provided state (status code 200)
     *         or The request was invalid or cannot be otherwise served. (status code 400)
     *         or An unexpected error occurred on the server. (status code 500)
     */
    @Operation(
        operationId = "listDevicesByState",
        summary = "List devices by state",
        description = "Convenience endpoint to list all devices in a given state.",
        tags = { "Devices" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Devices matching the provided state", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = DeviceDto.class)))
            }),
            @ApiResponse(responseCode = "400", description = "The request was invalid or cannot be otherwise served.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"BAD_REQUEST\",\"message\":\"Field 'name' is required.\"}"
                    )
                })

            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred on the server.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDto.class), examples = {
                    @ExampleObject(
                        name = "",
                        value = "{\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"Unexpected error. Please try again later.\"}"
                    )
                })

            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = DevicesApi.PATH_LIST_DEVICES_BY_STATE,
        produces = { "application/json" }
    )
    ResponseEntity<List<DeviceDto>> listDevicesByState(
        @NotNull @Parameter(name = "state", description = "Device state to filter by", required = true, in = ParameterIn.PATH) @PathVariable("state") StateDto state
    );

}