# target/generated-sources/openapi/src/main/java/com/company/device/device/autogenerated/dto/DeviceCreateRequestDto.java
package com.company.device.device.autogenerated.dto;

import java.net.URI;
import java.util.Objects;
import com.company.device.device.autogenerated.dto.StateDto;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import org.springframework.lang.Nullable;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * DeviceCreateRequestDto
 */

@JsonTypeName("DeviceCreateRequest")
@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-22T17:33:26.712568800+01:00[Europe/Warsaw]", comments = "Generator version: 7.17.0")
public class DeviceCreateRequestDto {

  private String name;

  private String brand;

  private StateDto state;

  public DeviceCreateRequestDto() {
    super();
  }

  /**
   * Constructor with only required parameters
   */
  public DeviceCreateRequestDto(String name, String brand, StateDto state) {
    this.name = name;
    this.brand = brand;
    this.state = state;
  }

  public DeviceCreateRequestDto name(String name) {
    this.name = name;
    return this;
  }

  /**
   * Get name
   * @return name
   */
  @NotNull @Size(min = 1, max = 200) 
  @Schema(name = "name", example = "Thermostat-7", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("name")
  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public DeviceCreateRequestDto brand(String brand) {
    this.brand = brand;
    return this;
  }

  /**
   * Get brand
   * @return brand
   */
  @NotNull @Size(min = 1, max = 200) 
  @Schema(name = "brand", example = "Pizer", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("brand")
  public String getBrand() {
    return brand;
  }

  public void setBrand(String brand) {
    this.brand = brand;
  }

  public DeviceCreateRequestDto state(StateDto state) {
    this.state = state;
    return this;
  }

  /**
   * Get state
   * @return state
   */
  @NotNull @Valid 
  @Schema(name = "state", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("state")
  public StateDto getState() {
    return state;
  }

  public void setState(StateDto state) {
    this.state = state;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DeviceCreateRequestDto deviceCreateRequest = (DeviceCreateRequestDto) o;
    return Objects.equals(this.name, deviceCreateRequest.name) &&
        Objects.equals(this.brand, deviceCreateRequest.brand) &&
        Objects.equals(this.state, deviceCreateRequest.state);
  }

  @Override
  public int hashCode() {
    return Objects.hash(name, brand, state);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DeviceCreateRequestDto {\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    brand: ").append(toIndentedString(brand)).append("\n");
    sb.append("    state: ").append(toIndentedString(state)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}