# target/generated-sources/openapi/src/main/java/com/company/device/device/autogenerated/dto/DeviceDto.java
package com.company.device.device.autogenerated.dto;

import java.net.URI;
import java.util.Objects;
import com.company.device.device.autogenerated.dto.StateDto;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import org.springframework.lang.Nullable;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

/**
 * DeviceDto
 */

@JsonTypeName("Device")
@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-22T17:33:26.712568800+01:00[Europe/Warsaw]", comments = "Generator version: 7.17.0")
public class DeviceDto {

  private Long id;

  private String name;

  private String brand;

  private StateDto state;

  public DeviceDto() {
    super();
  }

  /**
   * Constructor with only required parameters
   */
  public DeviceDto(Long id, String name, String brand, StateDto state) {
    this.id = id;
    this.name = name;
    this.brand = brand;
    this.state = state;
  }

  public DeviceDto id(Long id) {
    this.id = id;
    return this;
  }

  /**
   * Unique identifier for the device
   * @return id
   */
  @NotNull 
  @Schema(name = "id", example = "123456", description = "Unique identifier for the device", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("id")
  public Long getId() {
    return id;
  }

  public void setId(Long id) {
    this.id = id;
  }

  public DeviceDto name(String name) {
    this.name = name;
    return this;
  }

  /**
   * Human-friendly device name
   * @return name
   */
  @NotNull @Size(min = 1, max = 200) 
  @Schema(name = "name", example = "Gateway-01", description = "Human-friendly device name", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("name")
  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public DeviceDto brand(String brand) {
    this.brand = brand;
    return this;
  }

  /**
   * Manufacturer or brand of the device
   * @return brand
   */
  @NotNull @Size(min = 0, max = 200) 
  @Schema(name = "brand", example = "Acme", description = "Manufacturer or brand of the device", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("brand")
  public String getBrand() {
    return brand;
  }

  public void setBrand(String brand) {
    this.brand = brand;
  }

  public DeviceDto state(StateDto state) {
    this.state = state;
    return this;
  }

  /**
   * Get state
   * @return state
   */
  @NotNull @Valid 
  @Schema(name = "state", requiredMode = Schema.RequiredMode.REQUIRED)
  @JsonProperty("state")
  public StateDto getState() {
    return state;
  }

  public void setState(StateDto state) {
    this.state = state;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DeviceDto device = (DeviceDto) o;
    return Objects.equals(this.id, device.id) &&
        Objects.equals(this.name, device.name) &&
        Objects.equals(this.brand, device.brand) &&
        Objects.equals(this.state, device.state);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, name, brand, state);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DeviceDto {\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    brand: ").append(toIndentedString(brand)).append("\n");
    sb.append("    state: ").append(toIndentedString(state)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}