# target/generated-sources/openapi/src/main/java/com/company/device/device/autogenerated/dto/StateDto.java
package com.company.device.device.autogenerated.dto;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonValue;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

/**
 * Device lifecycle state
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-22T17:33:26.712568800+01:00[Europe/Warsaw]", comments = "Generator version: 7.17.0")
public enum StateDto {
  
  IN_USE("in-use"),
  
  AVAILABLE("available"),
  
  INACTIVE("inactive");

  private final String value;

  StateDto(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static StateDto fromValue(String value) {
    for (StateDto b : StateDto.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}